Refactored/Renamed ProductCard to be used in RI and YO, and to accept any product. Added "Add Outfit" functionality (though doesn't work right at all yet).
