+++ src/components/related_items_outfit/RelatedItems.jsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import PropTypes from 'prop-types';
 import { nanoid } from 'nanoid';
 import Carousel from './Carousel';
-import ProductCard from './RIProductCard';
+import ProductCard from './ProductCard';
 import Compare from './CompareModal';
 import sampleItems from './sampleRelatedItems';
 import sampleStyles from './sampleStylesData';
@@ -17,10 +17,7 @@ export default function RelatedItems() {
         <Carousel>
           {sampleItems.map((product) => (
             <ProductCard
-              category={product.category}
-              name={product.name}
-              price={parseFloat(product.default_price)}
-              features={product.features}
+              product={product}
               setOpenModal={setOpenModal}
               image={sampleStyles.results[0].photos[0].url}
               key={nanoid()}
